Remove extra comma from doc fixing table to break
Even though the comma on line 40 was escaped, the asciidoc render being used on git doesn't seem to respect the escape. There are visible escape characters in other parts of the doc too, but I gave up trying to fix. The table was completely broken though, and rather than reformatting the entire table in PSV or the like, I just removed the comma.